+++ packages/create-sitecore-jss/src/templates/angular/scripts/bootstrap.ts
@@ -1,8 +1,4 @@
 import 'dotenv/config';
-import { constants } from '@sitecore-jss/sitecore-jss-angular/cjs';
-import { generateConfig } from './generate-config';
-const projects = require('../angular.json').projects;
-const chalk = require('chalk');
 
 /*
   BOOTSTRAPPING
@@ -11,30 +7,15 @@ const chalk = require('chalk');
   and the global config module.
 */
 
-const disconnected = process.env.JSS_MODE === constants.JSS_MODE.DISCONNECTED;
-
-if (disconnected && process.env.FETCH_WITH === constants.FETCH_WITH.GRAPHQL) {
-  throw new Error(chalk.red("GraphQL requests to Dictionary and Layout services are not supported in disconnected mode."))
-}
+/*
+   PLUGINS GENERATION
+*/
+require('./generate-plugins');
 
 /*
   CONFIG GENERATION
-  Generates the /src/environments/environment.js file which contains runtime configuration
-  that the app can import and use.
-
-  This is generated rather than using Angular environments because of the need to set config params
-  based on build arguments, which env files don't directly allow.
 */
-function writeConfig(configOverride: { production: boolean, sitecoreApiHost?: string }, outputPath?: string) {
-  if (disconnected) {
-    configOverride.sitecoreApiHost = `http://localhost:${projects['<%- appName %>'].architect.serve.options.port}`;
-  }
-
-  generateConfig(configOverride, outputPath);
-}
-
-writeConfig({ production: false });
-writeConfig({ production: true }, 'src/environments/environment.prod.js');
+require('./generate-config');
 
 /*
   COMPONENT FACTORY GENERATION